requirements.txt
# ============================================================
# Multi-Agent AI Research Assistant — requirements.txt
# Python 3.11+ | Pinned for reproducibility
# ============================================================

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu

# --- Core LLM & Orchestration ---
langchain==0.3.25
langchain-core>=0.3.59,<1.0.0
langchain-community==0.3.24
langchain-groq==0.3.2
langgraph==0.4.1
langchain-anthropic==0.3.10

# --- LLM Providers ---
groq==0.23.1
anthropic==0.52.0                  # Fallback: claude-haiku

# --- Web Search ---
tavily-python==0.5.1

# --- Embeddings & Vector Store ---
sentence-transformers==3.4.1
faiss-cpu==1.9.0
# NOTE: Use faiss-gpu instead if you have a CUDA GPU available

# --- Document Loading ---
pypdf==5.4.0
requests==2.32.3
beautifulsoup4==4.13.4             # For URL scraping

# --- Backend ---
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.9.1           # For .env loading via BaseSettings
python-multipart==0.0.20           # For file uploads in FastAPI
httpx==0.28.1                      # Async HTTP client

# --- Frontend ---
streamlit==1.45.0
streamlit-autorefresh==1.0.1       # For live agent activity polling

# --- Evaluation ---
ragas==0.2.15
datasets==3.5.0                    # Required by RAGAS internals

# --- Observability & Logging ---
structlog==25.3.0                  # Structured JSON logging
tenacity==9.1.2                    # Retry logic with exponential backoff

# --- Utilities ---
python-dotenv==1.1.0
numpy==1.26.4
pandas==2.2.3
tiktoken==0.9.0                    # Token counting for context window management
tqdm==4.67.1

# --- Testing ---
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
httpx==0.28.1                      # Also used for async test client

# --- Dev / Formatting (optional but recommended) ---
black==25.1.0
ruff==0.11.5